api_requirements.txt
flask>=2.3.3
werkzeug>=2.3.7
ultralytics>=8.2.8
torch>=2.2.0
torchvision>=0.15.2
opencv-python>=4.8.0.76
numpy>=1.26.0
pillow>=10.3.0
